+++ src/entities/openai/api/get-market-opinion.ts
@@ -1,7 +1,7 @@
 import { httpClient } from '@/shared/api';
 import { AiAnalysisOpinion } from './schema';
 
-export const getOpenaiOpinion = async (): Promise<AiAnalysisOpinion> => {
+export const getMarketOpinion = async (): Promise<AiAnalysisOpinion> => {
   const response = await httpClient.post('/api/openai/opinion');
 
   return AiAnalysisOpinion.parse(response);
