+++ src/entities/openai/index.ts
@@ -1,0 +1,2 @@
+export { openaiQueries } from './api/query';
+export type { AiAnalysisOpinion } from './api/schema';
