+++ src/pages/home/api/query.ts
@@ -3,7 +3,6 @@ import {
   getStockRanks,
   type GetStockRanksRequestParams,
 } from './get-stock-ranks';
-import { getOpenaiOpinion } from './get-openai-opinion';
 
 export const rankQueries = {
   all: () => ['ranks', 'all'],
@@ -12,14 +11,5 @@ export const rankQueries = {
     queryOptions({
       queryKey: [...rankQueries.lists(), params],
       queryFn: () => getStockRanks(params),
-    }),
-};
-
-export const openaiQueries = {
-  all: () => ['openai', 'all'],
-  opinion: () =>
-    queryOptions({
-      queryKey: [...openaiQueries.all(), 'opinion'],
-      queryFn: () => getOpenaiOpinion(),
     }),
 };
