+++ src/pages/home/api/schema.ts
@@ -20,16 +20,3 @@ export const StockRank = z.object({
   bps: z.number(),
 });
 export type StockRank = z.infer<typeof StockRank>;
-
-export const AiAnalysisOpinion = z.object({
-  id: z.number(),
-  aiRequest: z.string(),
-  aiResponse: z.string(),
-  createdAt: z.string(),
-  stockCode: z.object({
-    stockCode: z.string(),
-    stockName: z.string(),
-    marketCategory: z.string(),
-  }),
-});
-export type AiAnalysisOpinion = z.infer<typeof AiAnalysisOpinion>;
